requirements.txt
pyserial==3.4
rpi.gpio==0.6.3